Refactor AdventyDistribuidorasHeader; move component and styles to new directory structure
